clean core module
