+++ core/eolearn/core/eonode.py
@@ -13,12 +13,12 @@ from dataclasses import dataclass, field
 from typing import Sequence, Optional, List, Tuple, Union, Set, Dict, cast
 
 from .eotask import EOTask
-from .utilities import generate_uid
+from .utilities.other import generate_uid
 
 
 @dataclass(frozen=True)
 class EONode:
-    """Class representing a node in EOWorkflow graph
+    """Class representing a node in EOWorkflow graph.
 
     The object id is kept to help with serialization issues. Tasks created in different sessions have a small chance
     of having an id clash. For this reason all tasks of a workflow should be created in the same session.
@@ -34,15 +34,15 @@ class EONode:
     uid: str = field(init=False, repr=False)
 
     def __post_init__(self):
-        """Additionally verifies the parameters and adds a unique id to the node"""
+        """Additionally verifies the parameters and adds a unique id to the node."""
         if not isinstance(self.task, EOTask):
-            raise ValueError(f"Value of `task` should be an instance of {EOTask.__name__}, got {self.task}")
+            raise ValueError(f"Value of `task` should be an instance of {EOTask.__name__}, got {self.task}.")
 
         if not isinstance(self.inputs, Sequence):
-            raise ValueError(f"Value of `inputs` should be a sequence (`list`, `tuple`, ...), got {self.inputs}")
+            raise ValueError(f"Value of `inputs` should be a sequence (`list`, `tuple`, ...), got {self.inputs}.")
         for input_node in self.inputs:
             if not isinstance(input_node, EONode):
-                raise ValueError(f"Values in `inputs` should be instances of {EONode.__name__}, got {input_node}")
+                raise ValueError(f"Values in `inputs` should be instances of {EONode.__name__}, got {input_node}.")
         super().__setattr__("inputs", tuple(self.inputs))
 
         if self.name is None:
@@ -51,13 +51,13 @@ class EONode:
         super().__setattr__("uid", generate_uid(self.task.__class__.__name__))
 
     def get_name(self, suffix_number: int = 0) -> str:
-        """Provides node name according to the class of the contained task and a given number"""
+        """Provides node name according to the class of the contained task and a given number."""
         if suffix_number:
             return f"{self.name}_{suffix_number}"
         return cast(str, self.name)
 
     def get_dependencies(self, *, _memo: Optional[Dict["EONode", Set["EONode"]]] = None) -> Set["EONode"]:
-        """Returns a set of nodes that this node depends on. Set includes the node itself"""
+        """Returns a set of nodes that this node depends on. Set includes the node itself."""
         _memo = _memo if _memo is not None else {}
         if self not in _memo:
             result = {self}.union(*(input_node.get_dependencies(_memo=_memo) for input_node in self.inputs))
@@ -91,7 +91,7 @@ def linearly_connect_tasks(*tasks: Union[EOTask, Tuple[EOTask, str]]) -> List[EO
 
 @dataclass(frozen=True)
 class NodeStats:
-    """An object containing statistical info about a node execution"""
+    """An object containing statistical info about a node execution."""
 
     node_uid: str
     node_name: str
