+++ core/eolearn/core/utilities/__init__.py
@@ -1,0 +1,3 @@
+"""
+A subfolder containing utilities
+"""
