+++ geometry/eolearn/tests/test_transformations.py
@@ -30,7 +30,7 @@ RASTER_FEATURE_TIMED = FeatureType.MASK, "RASTERIZED_CLM"
 
 
 CUSTOM_DATAFRAME = EOPatch.load(TEST_EOPATCH_PATH, lazy_loading=True)[VECTOR_FEATURE]
-CUSTOM_DATAFRAME = CUSTOM_DATAFRAME[(CUSTOM_DATAFRAME["AREA"] < 10**3)]
+CUSTOM_DATAFRAME = CUSTOM_DATAFRAME[(CUSTOM_DATAFRAME["AREA"] < 10 ** 3)]
 CUSTOM_DATAFRAME_3D = CUSTOM_DATAFRAME.copy()
 CUSTOM_DATAFRAME_3D.geometry = CUSTOM_DATAFRAME_3D.geometry.map(partial(shapely.ops.transform, lambda x, y: (x, y, 0)))
